+++ scripts/modifications16plus.js
@@ -9,7 +9,9 @@
 // Include modules
 let env = require('./env')
 let alert = require('./alert')
+let found = require('./found')
 let fs = require('fs-extra')
+let path = require('path')
 let rec = require('recursive-readdir')
 
 if (env.installed) {
@@ -25,6 +27,21 @@ if (env.installed) {
         fs.writeFileSync(file, content)
       }
     })
+    // Add missing language folder
+    if (found(env.app) && !found(env.app, 'lang')) {
+      fs.readdir(env.app, function (err, files) {
+        if (err) alert('Failed to add language folder modification.', 'issue')
+        if (files.length > 1) {
+          try {
+            fs.ensureDirSync(path.resolve(env.app, 'lang'))
+            fs.writeJsonSync(path.resolve(env.app, 'lang/' + (env.cfg.language || env.cfg.defaultLanguage || 'en') + '.json'), {}, { spaces: 2 })
+          } catch (err) {
+            alert('Failed to create languages folder.', 'issue')
+          }
+        }
+      })
+    }
+    // Alert
     alert('Release modifications of v1.6+ done.')
   })
 }